Add file info to span during tokenizing and handle error properly
