+++ crates/core/tests/compiling.rs
@@ -40,7 +40,7 @@ const SOURCE: &str = r#"
 fn compiles_constructor_bytecode() {
     // Lex and Parse the source code
     let flattened_source = FullFileSource { source: SOURCE, file: None, spans: vec![] };
-    let lexer = Lexer::new(flattened_source.source);
+    let lexer = Lexer::new(flattened_source.source, None);
     let tokens = lexer.into_iter().map(|x| x.unwrap()).collect::<Vec<Token>>();
     let mut parser = Parser::new(tokens, None);
 
@@ -66,7 +66,7 @@ fn compiles_constructor_bytecode() {
 fn compiles_runtime_bytecode() {
     // Lex and Parse the source code
     let flattened_source = FullFileSource { source: SOURCE, file: None, spans: vec![] };
-    let lexer = Lexer::new(flattened_source.source);
+    let lexer = Lexer::new(flattened_source.source, None);
     let tokens = lexer.into_iter().map(|x| x.unwrap()).collect::<Vec<Token>>();
     let mut parser = Parser::new(tokens, None);
