+++ crates/lexer/tests/builtins.rs
@@ -16,7 +16,7 @@ fn parses_builtin_function_in_macro_body() {
             "{", "}",
         );
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = Lexer::new(flattened_source.source);
+        let mut lexer = Lexer::new(flattened_source.source, None);
 
         let _ = lexer.next(); // whitespace
         let _ = lexer.next(); // #define
@@ -70,7 +70,7 @@ fn fails_to_parse_builtin_outside_macro_body() {
     for builtin in builtin_funcs {
         let source = &format!("{builtin}(MAIN)");
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = Lexer::new(flattened_source.source);
+        let mut lexer = Lexer::new(flattened_source.source, None);
 
         let tok = lexer.next();
         let unwrapped = tok.unwrap().unwrap();
@@ -101,7 +101,7 @@ fn fails_to_parse_invalid_builtin() {
             "{", "}",
         );
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = Lexer::new(flattened_source.source);
+        let mut lexer = Lexer::new(flattened_source.source, None);
 
         let _ = lexer.next(); // whitespace
         let _ = lexer.next(); // #define
