+++ crates/lexer/tests/context.rs
@@ -6,7 +6,7 @@ use huff_neo_utils::prelude::*;
 fn function_context() {
     let source = "#define function test(bytes32) {} returns (address)";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let lexer = Lexer::new(flattened_source.source);
+    let lexer = Lexer::new(flattened_source.source, None);
     let tokens = lexer.into_iter().map(|x| x.unwrap()).filter(|x| !matches!(x.kind, TokenKind::Whitespace)).collect::<Vec<Token>>();
 
     // check input
@@ -19,7 +19,7 @@ fn function_context() {
 fn event_context() {
     let source = "#define event Transfer(bytes32,address)";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let lexer = Lexer::new(flattened_source.source);
+    let lexer = Lexer::new(flattened_source.source, None);
     let tokens = lexer.into_iter().map(|x| x.unwrap()).filter(|x| !matches!(x.kind, TokenKind::Whitespace)).collect::<Vec<Token>>();
 
     assert_eq!(tokens.get(tokens.len() - 5).unwrap().kind, TokenKind::PrimitiveType(PrimitiveEVMType::Bytes(32)));
@@ -30,7 +30,7 @@ fn event_context() {
 fn macro_context() {
     let source = "#define macro TEST() = takes (0) returns (0) {byte}";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let lexer = Lexer::new(flattened_source.source);
+    let lexer = Lexer::new(flattened_source.source, None);
     let tokens = lexer.into_iter().map(|x| x.unwrap()).filter(|x| !matches!(x.kind, TokenKind::Whitespace)).collect::<Vec<Token>>();
     assert_eq!(tokens.get(tokens.len() - 3).unwrap().kind, TokenKind::Opcode(Opcode::Byte));
 }
