+++ crates/lexer/tests/eof.rs
@@ -5,7 +5,7 @@ use huff_neo_utils::prelude::*;
 fn end_of_file() {
     let source = " ";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = Lexer::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source, None);
 
     // Eats the whitespace
     let _ = lexer.next();
