+++ crates/lexer/tests/function_type.rs
@@ -9,7 +9,7 @@ fn parses_function_type() {
     for (fn_type, fn_type_kind) in fn_types {
         let source = &format!("#define function test() {fn_type} returns (uint256)");
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = Lexer::new(flattened_source.source);
+        let mut lexer = Lexer::new(flattened_source.source, None);
 
         let _ = lexer.next(); // #define
         let _ = lexer.next(); // whitespace
