+++ crates/lexer/tests/numbers.rs
@@ -5,7 +5,7 @@ use huff_neo_utils::prelude::*;
 fn lexes_zero_prefixed_numbers() {
     let source = "00";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = Lexer::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source, None);
 
     // The first and only token should be lexed as 0
     let tok = lexer.next().unwrap().unwrap();
@@ -21,7 +21,7 @@ fn lexes_zero_prefixed_numbers() {
 fn lexes_large_numbers() {
     let source = &format!("{}", usize::MAX);
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = Lexer::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source, None);
 
     // The first and only token should be lexed
     let tok = lexer.next().unwrap().unwrap();
