+++ crates/lexer/tests/symbols.rs
@@ -5,7 +5,7 @@ use huff_neo_utils::prelude::*;
 fn lexes_assign_op() {
     let source = "#define constant TRANSFER_EVENT_SIGNATURE =";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = Lexer::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source, None);
 
     // This token should be a Define identifier
     let tok = lexer.next();
@@ -58,7 +58,7 @@ fn lexes_assign_op() {
 fn lexes_brackets() {
     let source = "[TOTAL_SUPPLY_LOCATION] sload";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = Lexer::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source, None);
 
     // This token should be the open bracket
     let tok = lexer.next();
@@ -98,7 +98,7 @@ fn lexes_braces() {
     "#;
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
 
-    let mut lexer = Lexer::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source, None);
 
     // Eat the non-brace tokens
     let _ = lexer.next(); // whitespace
@@ -157,7 +157,7 @@ fn lexes_math_ops() {
     // MATHS
     let source = r#"100 + 10 - 20 * 5 / 4"#;
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = Lexer::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source, None);
 
     // Eat the number and whitespace
     let _ = lexer.next();
@@ -215,7 +215,7 @@ fn lexes_math_ops() {
 fn lexes_commas() {
     let source = "test,test";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = Lexer::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source, None);
     // Eat alphanumerics
     let _ = lexer.next();
 
@@ -237,7 +237,7 @@ fn lexes_commas() {
 fn lexes_comma_sparse() {
     let source = "test , test";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = Lexer::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source, None);
 
     let _ = lexer.next(); // alphanumerics
     let _ = lexer.next(); // whitespace
