+++ crates/parser/tests/abi.rs
@@ -7,7 +7,7 @@ fn build_abi_from_ast() {
     let source = "#define function test(uint256[2][],string) view returns(uint256)";
 
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let lexer = Lexer::new(flattened_source.source);
+    let lexer = Lexer::new(flattened_source.source, None);
     let tokens = lexer.into_iter().map(|x| x.unwrap()).filter(|x| !matches!(x.kind, TokenKind::Whitespace)).collect::<Vec<Token>>();
     let mut parser = Parser::new(tokens, None);
     let contract = parser.parse().unwrap();
